Using Classification Algorithms
By inspecting the data, we understand that this is a typical binary classification problem wherein we predicted sentiments associated with a particular review.

Positive Review: 1
Negative Review: 0

The only problem is that the input data (feature matrix) is not continuous or categorical in nature. It is actually textual.
In order to preprocess the data so that a machine learning algorithm can make sense out of it, we applied certain NLP techniques.